backend/requirements.txt
fastapi
uvicorn
pyyaml
bcrypt
PyJWT
pypinyin
